Change :project/artifactid to :project/name
Leiningen calls it name, in contrast to Maven where it's the artifactId.
Fix this in the :aliases entries in the README and in project.clj. The
code doesn't care, because it's a positional argument and not a named
one. Leave it "artifactid" there and just add a note. It's much better
than "name" anyway.